# Improve WeakList<> performance in degenerate scenarios