--- a/osu.Framework/Lists/IWeakList.cs
+++ b/osu.Framework/Lists/IWeakList.cs
@@ -21,14 +21,15 @@ namespace osu.Framework.Lists
         /// <summary>
         /// Adds a weak reference to this list.
         /// </summary>
-        /// <param name="weakReference"></param>
+        /// <param name="weakReference">The weak reference to add.</param>
         void Add(WeakReference<T> weakReference);
 
         /// <summary>
         /// Removes an item from this list.
         /// </summary>
-        /// <param name="item"></param>
-        void Remove(T item);
+        /// <param name="item">The item to remove.</param>
+        /// <returns>Whether the item was removed.</returns>
+        bool Remove(T item);
 
         /// <summary>
         /// Removes a weak reference from this list.