--- a/osu.Framework/Lists/LockedWeakList.cs
+++ b/osu.Framework/Lists/LockedWeakList.cs
@@ -28,10 +28,10 @@ namespace osu.Framework.Lists
                 list.Add(weakReference);
         }
 
-        public void Remove(T item)
+        public bool Remove(T item)
         {
             lock (list)
-                list.Remove(item);
+                return list.Remove(item);
         }
 
         public bool Remove(WeakReference<T> weakReference)