@peppy 2020-05-19
The target for optimisation here is `RemoveAllIteratively` with many items being removed, correct?