@smoogipoo 2020-05-19
That's correct, yes.