@bdach 2020-05-23
Game-side CI has exposed a possible [failure scenario](https://ci.appveyor.com/project/peppy/osu/builds/33061305) that seems related to these changes. Not sure whether it needs fixing framework-side, or it's a case of unsafe `WeakList` usage as I have not investigated in detail yet.